requirements.txt
absl-py==0.9.0
asgiref==3.2.3
astor==0.8.1
astunparse==1.6.3
autopep8==1.5.2
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.6
django-cors-headers==3.2.1
djangorestframework==3.11.0
elasticsearch==5.5.3
gast==0.3.3
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gprof2dot==2019.11.30
grpcio==1.28.1
h5py==2.10.0
idna==2.8
importlib-metadata==1.6.0
Jinja2==2.11.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.2.1
panda==0.3.1
protobuf==3.11.3
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.13.0
sqlparse==0.3.0
tb-nightly==1.15.0a20190821
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019090401
urllib3==1.25.7
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0

